pyproject.toml
[tool.poetry]
name = "django-easy-plants"
version = "0.1.0"
description = ""
authors = ["Kai Diefenbach <kai.diefenbach@iqpp.de>"]
readme = "README.md"
packages = [{include = "easy_plants", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
